+++ CHANGES.rst
@@ -72,6 +72,7 @@ Fixed:
 * Items inside a table cannot be referenced (issue #174)
 * Sphinx frontend: fix handling of relative image paths in .rst files inside
   a directory in the Sphinx project root
+* rinoh: fix --install-resources (broken since PyPI disabled XMLRPC searches)
 * GroupedLabeledFlowables: respect label_min_width and fix a crash with respect
   to space_below handling
 * Duplicate rendering of content in columns; if content was too small to fill